updated the documentation and the npm scripts
